+++ src/reader/_plugins/preview_feed_list.py
@@ -5,12 +5,12 @@ preview_feed_list
 If the feed to be previewed is not actually a feed,
 show a list of feeds linked from that URL (if any).
 
-This plugin needs additional dependencies, use the ``preview-feed-list`` extra
+This plugin needs additional dependencies, use the ``unstable-plugins`` extra
 to install them:
 
 .. code-block:: bash
 
-    pip install reader[preview-feed-list]
+    pip install reader[unstable-plugins]
 
 To load::